update the seeded scripts so that old types aren't in use
