+++ scripts/grammys.json
@@ -1,5 +1,5 @@
 {
-  "name": "Grammys 2019",
+  "name": "62nd Grammy Awards",
   "categories": [
     {
       "name": "Best New Artist",
